Safari-609.1.11.1/LayoutTests/js/arraybuffer-wrappers-expected.txt

Tests wrappers for ArrayBuffer objects are not GCed when they shouldn't be

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test subclassing
PASS foo.buffer instanceof C is true
PASS foo.buffer instanceof C is true
PASS foo.buffer instanceof C is true
PASS foo.buffer instanceof C is true
PASS foo.buffer instanceof C is true
PASS foo.buffer instanceof C is true
PASS foo.buffer instanceof C is true
PASS foo.buffer instanceof C is true
PASS foo.buffer instanceof C is true
PASS foo.buffer instanceof C is true

Test properties
PASS foo.buffer.bar is 1
PASS foo.buffer.bar is 1
PASS foo.buffer.bar is 1
PASS foo.buffer.bar is 1
PASS foo.buffer.bar is 1
PASS foo.buffer.bar is 1
PASS foo.buffer.bar is 1
PASS foo.buffer.bar is 1
PASS foo.buffer.bar is 1
PASS foo.buffer.bar is 1

Test WeakMap
PASS ws.has(foo.buffer) is true
PASS ws.has(foo.buffer) is true
PASS ws.has(foo.buffer) is true
PASS ws.has(foo.buffer) is true
PASS ws.has(foo.buffer) is true
PASS ws.has(foo.buffer) is true
PASS ws.has(foo.buffer) is true
PASS ws.has(foo.buffer) is true
PASS ws.has(foo.buffer) is true
PASS ws.has(foo.buffer) is true
PASS successfullyParsed is true

TEST COMPLETE
